chore(rollup): Updates the rollup configuration for all libraries to match the template.
