+++ lib/eslint-plugin-reactotron/rollup.config.cjs
@@ -5,7 +5,7 @@ import typescript from "@rollup/plugin-typescript"
 const pkg = require("./package.json")
 
 export default {
-  input: "src/index.ts",
+  input: "src/eslint-plugin-reactotron.ts",
   output: [
     {
       file: pkg.main,
