+++ lib/reactotron-apisauce/rollup.config.cjs
@@ -7,7 +7,7 @@ const pkg = require("./package.json")
 
 /** @type {import('rollup').RollupOptions} */
 export default {
-  input: "src/index.ts",
+  input: "src/reactotron-apisauce.ts",
   output: [
     {
       file: pkg.main,
