+++ lib/reactotron-core-ui/rollup.config.cjs
@@ -5,7 +5,7 @@ import typescript from "@rollup/plugin-typescript"
 const pkg = require("./package.json")
 
 export default {
-  input: "src/index.ts",
+  input: "src/reactotron-core-ui.ts",
   output: [
     {
       file: pkg.main,
