+++ lib/reactotron-react-js/rollup.config.cjs
@@ -5,7 +5,7 @@ import typescript from "@rollup/plugin-typescript"
 const pkg = require("./package.json")
 
 export default {
-  input: "src/index.ts",
+  input: "src/reactotron-react-js.ts",
   output: [
     {
       file: pkg.main,
