+++ lib/reactotron-react-js/src/index.ts
@@ -1,1 +1,3 @@
+import reactJsClient from "./reactotron-react-js"
 export * from "./reactotron-react-js"
+export default reactJsClient
