+++ lib/reactotron-react-native/src/index.ts
@@ -1,3 +1,3 @@
-import reactotron from "./reactotron-react-native"
+import reactotronReactNative from "./reactotron-react-native"
 export * from "./reactotron-react-native"
-export default reactotron
+export default reactotronReactNative
