+++ lib/reactotron-redux/rollup.config.cjs
@@ -4,7 +4,7 @@ import typescript from "@rollup/plugin-typescript"
 const pkg = require("./package.json")
 
 export default {
-  input: "src/index.ts",
+  input: "src/reactotron-redux.ts",
   output: [
     {
       file: pkg.main,
